Speed Optimizations
I made a bunch of changes, mostly making both MyStack and MyQueue into sub-classes of MyLinkedList. This led to instead of using the methods in MyLinkedList on MyStack and MyQueue to write push, pop, queue and dequeue, just writing out original code for all of them. I also included a tail for MyQueue so that it would run without using the traversals (they're still there if need be) in order to save runtime.

tl;dr: I did a lot of work to decrease runtime. :+1: