Aight so about it working
So I realized if I want to use my graph as a base for my tree, I would have s *smidge* of an issue when it came to traversals due to my traversals being based on the value of the item as an integer. So I made the whole graph able to be generic typed. So thats a thing now.